+++ changelog.md
@@ -9,6 +9,8 @@
   * Min and max attack distance is now configurable
   * Slowness is no longer applied from thrown webs if the player parrys the projectile
 * Launching + Breaching creepers will now launch and explode on contact with walls
+* Slightly adjusted throwing web and skeleton arrows shooting angle
+* Fixed skeletons not taking into account config inaccuracy
 * Fixed breaching only working when creepers were in water
 * Fixed spiders throwing webs even they could not see the player
 * Fixed skeletons running from target having faster speed when far away instead of when near. Also made those speeds configurable.